move demandsSection.svelte to shared components
