+++ src/Home/Home.svelte
@@ -1,6 +1,6 @@
 <script>
   import HomeHeader from './HomeHeader/HomeHeader.svelte'
-  import DemandsSection from './DemandsSection/DemandsSection.svelte'
+  import DemandsSection from '../sharedComponents/DemandsSection.svelte'
   import ChangeSection from './ChangeSection/ChangeSection.svelte'
   import IntersectionalitySection from './IntersectionalitySection/IntersectionalitySection.svelte'
   import WeVotedSection from './WeVotedSection/WeVotedSection.svelte'
